move equations out of equations
